admin de inventory lvl ok liao gua and also fix the product delete will oso delete specification of the product

+++ admin/product_delete.php
@@ -5,18 +5,34 @@ include '../admin_head.php';
 if (is_post()) {
     $ProductID = req('ProductID');
 
-    // Delete photo
+    // Delete photo thumbnail photo
     $stm = $_db->prepare('SELECT ProductThumb FROM product WHERE ProductID = ?');
     $stm->execute([$ProductID]);
-    $photo = $stm->fetchColumn();
-
+    $thumb = $stm->fetchColumn();
     // Remove photo from the physical folder
-    unlink("../images/product/$photo");
+    if ($thumb && file_exists("../images/product/$thumb")) {
+        unlink("../images/product/$thumb");
+    }
 
-    // Delete record from DB
+    // Delete each specification photo linked to this product
+    $stm = $_db->prepare('SELECT ProductPhoto FROM specification WHERE ProductID = ?');
+    $stm->execute([$ProductID]);
+    $photos = $stm->fetchAll(PDO::FETCH_COLUMN);
+    foreach ($photos as $photo) {
+        if ($photo && file_exists("../images/product/$photo")) {
+            unlink("../images/product/$photo");
+        }
+    }
+
+    // Delete specification records for this product
+    $stm = $_db->prepare('DELETE FROM specification WHERE ProductID = ?');
+    $stm->execute([$ProductID]);
+
+    // Delete product record from DB
     $stm = $_db->prepare('DELETE FROM product WHERE ProductID = ?');
     $stm->execute([$ProductID]);
-    temp('info', 'Product record deleted');
+
+    temp('info', 'Product and its specifications deleted successfully');
 }
 
 redirect('product.php');
